Move supplier_data helper to supplier_data_helpers
